web-app/requirements.txt
streamlit
Pillow
boto3